feat(works): cria e refator DTOs para a Watchlist

+++ src/main/java/com/example/api_ecw/works/dto/TmdbWorkResponse.java
@@ -3,11 +3,10 @@ package com.example.api_ecw.works.dto;
 import com.fasterxml.jackson.annotation.JsonAlias;
 import com.fasterxml.jackson.annotation.JsonProperty;
 
-import java.lang.reflect.Array;
 import java.util.List;
 
 public record TmdbWorkResponse(
-        Long id,
+        Integer id,
         @JsonProperty("original_title") @JsonAlias("original_name") String originalTitle,
         @JsonProperty("title") @JsonAlias("name") String title,
         @JsonProperty("original_language") String originalLanguage,
